+++ src/utils/__tests__/selections.spec.js
@@ -6,7 +6,6 @@ describe('selections', () => {
     const { select } = selections;
     let node;
     let selectionObj;
-    let selectionState;
     let isActive;
     beforeEach(() => {
       node = JSON.parse(JSON.stringify(defaultValues.nodes));
@@ -20,51 +19,51 @@ describe('selections', () => {
         },
         linked: false,
         setState: sinon.spy(),
+        state: [],
       };
-      selectionState = [];
     });
     it('should early return if elemNo is negative', () => {
       node.data.elemNo = -2;
-      select(node, selectionObj, selectionState);
+      select(node, selectionObj);
       expect(selectionObj.setState).to.not.have.been.called;
     });
     it('should early return if elemNo isLocked is true', () => {
       node.data.isLocked = true;
-      select(node, selectionObj, selectionState);
+      select(node, selectionObj);
       expect(selectionObj.setState).to.not.have.been.called;
     });
     it('should call begin and setState to node id', () => {
-      select(node, selectionObj, selectionState);
+      select(node, selectionObj);
       expect(selectionObj.api.begin).to.have.been.calledOnce;
       expect(selectionObj.api.select).to.have.been.calledOnce;
       expect(selectionObj.setState).to.have.been.calledWith([1]);
     });
     it('should not call begin and push node id to selectionState', () => {
       isActive = true;
-      selectionState.push(2);
-      select(node, selectionObj, selectionState);
+      selectionObj.state.push(2);
+      select(node, selectionObj);
       expect(selectionObj.api.begin).to.not.have.been.called;
       expect(selectionObj.api.select).to.have.been.calledOnce;
       expect(selectionObj.setState).to.have.been.calledWith([2, 1]);
     });
     it('should get all children when linked is true', () => {
       selectionObj.linked = true;
-      select(node, selectionObj, selectionState);
+      select(node, selectionObj);
       expect(selectionObj.api.select).to.have.been.calledOnce;
       expect(selectionObj.setState).to.have.been.calledWith([1, 2, 3, 798, 88]);
     });
 
     it('should deselect node', () => {
       isActive = true;
-      selectionState = [1, 2];
-      select(node, selectionObj, selectionState);
+      selectionObj.state = [1, 2];
+      select(node, selectionObj);
       expect(selectionObj.setState).to.have.been.calledWith([2]);
     });
     it('should deselect node and children', () => {
       isActive = true;
       selectionObj.linked = true;
-      selectionState = [1, 2];
-      select(node, selectionObj, selectionState);
+      selectionObj.state = [1, 2];
+      select(node, selectionObj);
       expect(selectionObj.api.clear).to.have.been.calledOnce;
       expect(selectionObj.setState).to.have.been.calledWith([]);
     });
